➕ qt 6.8.2 (#21)
Of course docker being docker, when it's been a while it's not the same system anymore, so aqt wasn't compatible anymore with the python version in ubuntu 20.
Easy workaround is to create an intermediate build step on ubuntu 24 to install aqt (venv mandatory in ubuntu 24, system wide installation is forbidden). Then this download the qt version we want.
The main build step copy the folder from the intermediary build step